+++ app/models/group.rb
@@ -2,7 +2,7 @@
 class Group < ApplicationRecord
   include CheckExistenceConcern
 
-  has_many :user_groups
+  has_many :user_groups, autosave: true
   has_many :users, through: :user_groups
   has_many :access_policies, dependent: :destroy
   has_many :parent_group_relationship,
@@ -26,9 +26,6 @@ class Group < ApplicationRecord
   validates :name, uniqueness: true
   validate :at_least_one_owner
 
-  before_validation :owner_ids_into_user_groups
-  before_validation :member_ids_into_user_groups
-
   def ancestors
     parents + parents.map(&:ancestors).flatten
   end
@@ -45,16 +42,6 @@ class Group < ApplicationRecord
     (users + offspring.map(&:users).flatten).uniq
   end
 
-  attr_writer :member_ids
-  def member_ids
-    members.pluck(:id)
-  end
-
-  attr_writer :owner_ids
-  def owner_ids
-    owners.pluck(:id)
-  end
-
   def members
     users.joins(:user_groups).where(user_groups: { owner: false }).distinct
   end
@@ -64,34 +51,6 @@ class Group < ApplicationRecord
   end
 
   private
-
-  def owner_ids_into_user_groups
-    return unless @owner_ids
-
-    owners = User.where(id: @owner_ids)
-    create_user_groups(owners, owner: true)
-    destroy_non_existing(owners, owner: true)
-  end
-
-  def member_ids_into_user_groups
-    return unless @member_ids
-
-    @member_ids -= @owner_ids if @owner_ids
-    members = User.where(id: @member_ids)
-    create_user_groups(members, owner: false)
-    destroy_non_existing(members, owner: false)
-  end
-
-  def create_user_groups(members, owner:)
-    members.each do |member|
-      user_group = user_groups.find_or_initialize_by(user: member)
-      user_group.owner = owner
-    end
-  end
-
-  def destroy_non_existing(members, owner:)
-    user_groups.where(owner: owner).where.not(user: members).destroy_all
-  end
 
   def at_least_one_owner
     errors.add(:owner_ids, 'At least one group owner is required') unless user_groups.any?(&:owner)