+++ app/models/user_group.rb
@@ -1,5 +1,5 @@
 # frozen_string_literal: true
 class UserGroup < ApplicationRecord
   belongs_to :user
-  belongs_to :group
+  belongs_to :group, autosave: true
 end